# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education Platform API",
        "description": "API documentation for the Education Platform. This API provides comprehensive endpoints for managing educational content including articles, users, and files.",
        "contact": {
            "name": "Support Team",
            "email": "support@alemedu.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/grade-one": {
            "get": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Get list of first grade content",
                "operationId": "d3a1a693d16627dd77ca41fb89d1d299",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "Filter by subject",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "description": "Filter by unit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GradeOne"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Create new first grade content",
                "operationId": "0ec4b82b5f78da65adaa541333069388",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradeOneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Content created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeOne"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grade-one/{id}": {
            "get": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Get first grade content by ID",
                "operationId": "543c1185fa7f9580bbb9131f230b4eb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Content ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeOne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Content not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Update first grade content",
                "operationId": "0e8602f0014daedacaef4822c5cdc120",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Content ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradeOneRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Content updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeOne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Content not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Delete first grade content",
                "operationId": "955f546c446d9b5c09e890e3748b6d75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Content ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content deleted successfully"
                    },
                    "404": {
                        "description": "Content not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grade-one/{id}/attachment": {
            "post": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Upload attachment for first grade content",
                "operationId": "651086be58a8ad23ec469f39974d4281",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Content ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "attachment": {
                                        "description": "Attachment file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attachment uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attachment": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grade-one/{id}/attachment/{filename}": {
            "delete": {
                "tags": [
                    "Grade One"
                ],
                "summary": "Delete attachment from first grade content",
                "operationId": "62c45ed3882a79c69a53ed7520953a43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Content ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Attachment filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attachment deleted successfully"
                    },
                    "404": {
                        "description": "Content or attachment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GradeOne": {
                "title": "Grade One",
                "description": "Grade One object schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Mathematics"
                    },
                    "description": {
                        "type": "string",
                        "example": "First grade mathematics curriculum"
                    },
                    "subject": {
                        "type": "string",
                        "example": "math"
                    },
                    "unit": {
                        "type": "string",
                        "example": "Numbers and Operations"
                    },
                    "lesson": {
                        "type": "string",
                        "example": "Addition"
                    },
                    "content": {
                        "type": "string",
                        "example": "Detailed lesson content..."
                    },
                    "video_url": {
                        "type": "string",
                        "example": "https://example.com/video.mp4"
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "GradeOneRequest": {
                "title": "Grade One Request",
                "description": "Grade One request schema",
                "required": [
                    "title",
                    "subject",
                    "unit",
                    "lesson",
                    "content"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Mathematics"
                    },
                    "description": {
                        "type": "string",
                        "example": "First grade mathematics curriculum"
                    },
                    "subject": {
                        "type": "string",
                        "example": "math"
                    },
                    "unit": {
                        "type": "string",
                        "example": "Numbers and Operations"
                    },
                    "lesson": {
                        "type": "string",
                        "example": "Addition"
                    },
                    "content": {
                        "type": "string",
                        "example": "Detailed lesson content..."
                    },
                    "video_url": {
                        "type": "string",
                        "example": "https://example.com/video.mp4"
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}